+++ internal/commands/root_test.go
@@ -48,6 +48,7 @@ func createASTTestCommand() *cobra.Command {
 	learnMoreMockWrapper := &mock.LearnMoreMockWrapper{}
 	prMockWrapper := &mock.PRMockWrapper{}
 	tenantConfigurationMockWrapper := &mock.TenantConfigurationMockWrapper{}
+	jwtWrapper := &mock.JWTMockWrapper{}
 
 	return NewAstCLI(
 		scansMockWrapper,
@@ -69,6 +70,7 @@ func createASTTestCommand() *cobra.Command {
 		prMockWrapper,
 		learnMoreMockWrapper,
 		tenantConfigurationMockWrapper,
+		jwtWrapper,
 	)
 }
